backend/requirements.txt
fastapi
uvicorn[standard]
zaproxy
openai
python-dotenv
pydantic<2
